# xx/api/api.bm.go
// Code generated by protoc-gen-bm v0.1, DO NOT EDIT.
// source: api2.proto

/*
Package api is a generated blademaster stub package.
This code was generated with kratos/tool/protobuf/protoc-gen-bm v0.1.

package 命名使用 {appid}.{version} 的方式, version 形如 v1, v2 ..

It is generated from these files:
	api2.proto
*/
package api

import (
	"context"

	bm "github.com/go-kratos/kratos/pkg/net/http/blademaster"
	"github.com/go-kratos/kratos/pkg/net/http/blademaster/binding"
)
import google_protobuf1 "google.golang.org/protobuf/types/known/emptypb"

// to suppressed 'imported but not used warning'
var _ *bm.Context
var _ context.Context
var _ binding.StructValidator

var PathJobPing = "/demo.service.v1.Job/Ping"
var PathJobSayHello = "/demo.service.v1.Job/SayHello"
var PathJobSayHelloURL = "/kratos-demo/say_hello"

// JobBMServer is the server API for Job service.
type JobBMServer interface {
	Ping(ctx context.Context, req *google_protobuf1.Empty) (resp *google_protobuf1.Empty, err error)

	SayHello(ctx context.Context, req *HelloReq) (resp *google_protobuf1.Empty, err error)

	SayHelloURL(ctx context.Context, req *HelloReq) (resp *HelloResp, err error)
}

var JobSvc JobBMServer

func jobPing(c *bm.Context) {
	p := new(google_protobuf1.Empty)
	if err := c.BindWith(p, binding.Default(c.Request.Method, c.Request.Header.Get("Content-Type"))); err != nil {
		return
	}
	resp, err := JobSvc.Ping(c, p)
	c.JSON(resp, err)
}

func jobSayHello(c *bm.Context) {
	p := new(HelloReq)
	if err := c.BindWith(p, binding.Default(c.Request.Method, c.Request.Header.Get("Content-Type"))); err != nil {
		return
	}
	resp, err := JobSvc.SayHello(c, p)
	c.JSON(resp, err)
}

func jobSayHelloURL(c *bm.Context) {
	p := new(HelloReq)
	if err := c.BindWith(p, binding.Default(c.Request.Method, c.Request.Header.Get("Content-Type"))); err != nil {
		return
	}
	resp, err := JobSvc.SayHelloURL(c, p)
	c.JSON(resp, err)
}

// RegisterJobBMServer Register the blademaster route
func RegisterJobBMServer(e *bm.Engine, server JobBMServer) {
	JobSvc = server
	e.GET("/demo.service.v1.Job/Ping", jobPing)
	e.GET("/demo.service.v1.Job/SayHello", jobSayHello)
	e.GET("/kratos-demo/say_hello", jobSayHelloURL)
}